+++ rider_information.py
@@ -1,4 +1,4 @@
-import driver_manager
+import appium_driver
 
 # device_info 리스트 순서에 맞춘 테스트 계정 정보 리스트
 test_accounts = [
@@ -15,7 +15,7 @@ test_accounts = [
 # 순서대로 udid -> 계정 정보 매핑 딕셔너리 생성
 device_test_data = {}
 
-for i, device in enumerate(driver_manager.device_info):
+for i, device in enumerate(appium_driver.device_info):
     if i < len(test_accounts):
         device_test_data[device["udid"]] = test_accounts[i]
     else:
